Reorganised routes, now more simple and better
